+++ src/data/dataset.cc
@@ -27,7 +27,7 @@
 namespace quickrank {
 namespace data {
 
-Dataset::Dataset(unsigned int n_instances, unsigned int n_features) {
+Dataset::Dataset(size_t n_instances, size_t n_features) {
   max_instances_ = n_instances;
   num_features_ = n_features;
   num_instances_ = 0;
@@ -77,7 +77,7 @@ void Dataset::addInstance(QueryID q_id, Label i_label,
   // update label and features
   labels_[num_instances_] = i_label;
   quickrank::Feature* new_instance = data_ + (num_instances_ * num_features_);
-  for (unsigned int i = 0; i < i_features.size(); i++)
+  for (size_t i = 0; i < i_features.size(); i++)
     new_instance[i] = i_features[i];
 
   // update offset of last query result
@@ -90,8 +90,8 @@ void Dataset::addInstance(QueryID q_id, Label i_label,
   offsets_.back() = num_instances_;
 }
 
-std::unique_ptr<QueryResults> Dataset::getQueryResults(unsigned int i) const {
-  unsigned int num_results = offsets_[i + 1] - offsets_[i];
+std::unique_ptr<QueryResults> Dataset::getQueryResults(size_t i) const {
+  size_t num_results = offsets_[i + 1] - offsets_[i];
   quickrank::Feature* start_data = data_
       + ((format_ == HORIZ) ? (offsets_[i] * num_features_) : (offsets_[i]));
   quickrank::Label* start_label = labels_ + offsets_[i];
@@ -112,8 +112,8 @@ void Dataset::transpose() {
     exit(EXIT_FAILURE);
   }
 
-  for (unsigned int i = 0; i < num_instances_; i++) {
-    for (unsigned int f = 0; f < num_features_; f++) {
+  for (size_t i = 0; i < num_instances_; i++) {
+    for (size_t f = 0; f < num_features_; f++) {
       if (format_ == HORIZ)
         transposed[f * num_instances_ + i] = data_[i * num_features_ + f];
       else
